+++ packages/rax-recyclerview/CHANGELOG.md
@@ -2,8 +2,9 @@
 
 ## 2.0.0
 
-- Refactor rax-recyclerview
-- upgrade deps: `rax-scrollview: ^3.6.0`
+- Refactor rax-recyclerview:
+  - 
+- upgrade deps: `rax-scrollview` from `^3.3.3` to `^3.6.0`
 
 ## 1.3.6